Admin

+++ app/Livewire/Client/CourseCreation/Components/CourseThumbnail.php
@@ -3,9 +3,7 @@
 namespace App\Livewire\Client\CourseCreation\Components;
 
 use Exception;
-use Illuminate\Contracts\View\Factory;
 use Illuminate\Contracts\View\View;
-use Illuminate\Foundation\Application;
 use Illuminate\Support\Facades\File;
 use Illuminate\Support\Facades\Storage;
 use Livewire\Attributes\Modelable;
@@ -20,17 +18,14 @@ class CourseThumbnail extends Component {
     public $thumbnail;
 
     public $image;
-    public string $imagePreview;
+    public string $imagePreview = '';
 
-    /**
-     * @throws Exception
-     */
     public function updatedImage(): void
     {
         try {
             $this->validate(['image' => 'image|mimes:jpeg,png,jpg,webp'], [
-                'image.image' => 'The file must be an image.',
-                'image.mimes' => 'The file must be a file of type: jpeg, png, jpg, webp.',
+                'image.image' => 'Tệp tải lên phải là một hình ảnh.',
+                'image.mimes' => 'Định dạng ảnh không hỗ trợ. Vui lòng dùng: jpeg, png, jpg, webp.',
             ]);
             $this->imagePreview = $this->image->temporaryUrl();
             $this->thumbnail = $this->image->getFilename();
@@ -52,17 +47,17 @@ class CourseThumbnail extends Component {
     #[On('course-creation-submitted')]
     public function storeThumbnail(): void
     {
-        if (Storage::disk('local')->exists('livewire-tmp/' . $this->thumbnail)) {
+        if ($this->thumbnail && Storage::disk('local')->exists('livewire-tmp/' . $this->thumbnail)) {
             $this->image->storeAs(
-                path: 'course/thumbnails',
-                options: 'public',
-                name: $this->thumbnail
+                path: config('filesystems.paths.course.internal.thumbnail.pending'),
+                name: $this->thumbnail,
+                options: 'public'
             );
             $this->deleteImage();
         }
     }
 
-    public function render(): Factory|Application|View
+    public function render(): View
     {
         return view('livewire.client.course-creation.components.course-thumbnail');
     }
